Add files via upload
RFM segmentation is a marketing technique that groups customers into different segments based on their recency, frequency, and monetary value. This information can then be used to target customers with relevant marketing messages and campaigns.

This project will implement RFM segmentation to analyze customer behavior and identify different customer segments. The project will use a Pandas Python library to perform the RFM segmentation and then give the results.

Although i worked on this project with my team lead who was extremely supportive to me. he helped me to make this project more optimized by using many techniques that i wasn't aware of such 'Chaining coding technique' and 'Writing comment to every block of code' so that it is easy to understand to anyone.
with this, here is the project!!